Move forum admin to /admin as well
