+++ app/Http/routes.php
@@ -130,10 +130,6 @@ Route::group(['prefix' => 'forum'], function () {
     Route::patch('p/{posts}', ['as' => 'forum.posts.update', 'uses' => "Forum\PostsController@update"]);
     Route::get('p/{posts}/edit', ['as' => 'forum.posts.edit', 'uses' => "Forum\PostsController@edit"]);
     Route::get('p/{posts}/raw', ['as' => 'forum.posts.raw', 'uses' => "Forum\PostsController@raw"]);
-
-    Route::group(['prefix' => 'admin', 'namespace' => 'Forum\Admin'], function () {
-        Route::resource('forum-covers', 'ForumCoversController', ['only' => ['index', 'store', 'update']]);
-    });
 });
 
 Route::group(['prefix' => 'admin', 'namespace' => 'Admin'], function () {
@@ -148,6 +144,10 @@ Route::group(['prefix' => 'admin', 'namespace' => 'Admin'], function () {
         Route::resource('addresses', 'AddressController', ['only' => ['update']]);
 
         Route::get('/', function () { return Redirect::route('admin.store.orders.index'); });
+    });
+
+    Route::group(['prefix' => 'forum', 'namespace' => 'Forum'], function () {
+        Route::resource('forum-covers', 'ForumCoversController', ['only' => ['index', 'store', 'update']]);
     });
 });
