+++ resources/lang/en/admin.php
@@ -1,0 +1,45 @@
+<?php
+
+/**
+ *    Copyright 2015 ppy Pty. Ltd.
+ *
+ *    This file is part of osu!web. osu!web is distributed in the hopes of
+ *    attracting more community contributions to the core ecosystem of osu!
+ *
+ *    osu!web is free software: you can redistribute it and/or modify
+ *    it under the terms of the Affero GNU General Public License version 3
+ *    as published by the Free Software Foundation.
+ *
+ *    osu!web is distributed WITHOUT ANY WARRANTY; without even the implied
+ *    warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *    See the GNU Affero General Public License for more details.
+ *
+ *    You should have received a copy of the GNU Affero General Public License
+ *    along with osu!web.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+return [
+
+    'forum' => [
+        'forum-covers' => [
+            'delete' => 'Delete',
+
+            'forum-name' => 'Forum #:id: :name',
+
+            'no-cover' => 'No cover set',
+
+            'submit' => [
+                'save' => 'Save',
+                'update' => 'Update',
+            ],
+
+            'title' => 'Forum Covers List',
+
+            'type-title' => [
+                'default-topic' => 'Default Topic Cover',
+                'main' => 'Forum Cover',
+            ],
+        ],
+    ],
+
+];
