+++ resources/views/admin/forum/forum_covers/index.blade.php
@@ -19,14 +19,14 @@
 
 @section("content")
     <div class="osu-layout__row osu-layout__row--page">
-        <h1>{{ trans('forum.admin.forum-covers.title') }}</h1>
+        <h1>{{ trans('admin.forum.forum-covers.title') }}</h1>
 
         @foreach ($forums as $forum)
             <div class="forum-cover-admin-item" id="forum-{{ $forum->forum_id }}">
                 <h2>
                     {!! link_to(
                         route('forum.forums.show', $forum->forum_id),
-                        trans('forum.admin.forum-covers.forum-name', ['id' => $forum->forum_id, 'name' => $forum->forum_name])
+                        trans('admin.forum.forum-covers.forum-name', ['id' => $forum->forum_id, 'name' => $forum->forum_name])
                     ) !!}
                 </h2>
 
@@ -35,7 +35,7 @@
                     'default-topic' => ['cover' => $forum->cover->defaultTopicCover ?? null, 'key' => 'default_topic_cover']
                 ] as $type => $cover)
                     <div class="forum-cover-admin-item__cover">
-                        <h3>{{ trans("forum.admin.forum-covers.type-title.{$type}") }}</h3>
+                        <h3>{{ trans("admin.forum.forum-covers.type-title.{$type}") }}</h3>
 
                         @if ($cover['cover'] !== null && $cover['cover']->fileUrl() !== null)
                             <div class="forum-cover-admin-item__cover-image"
@@ -43,7 +43,7 @@
                             ></div>
 
                             {!! Form::open([
-                                'route' => ['forum.admin.forum-covers.update', $cover['cover']->id],
+                                'route' => ['admin.forum.forum-covers.update', $cover['cover']->id],
                                 'method' => 'POST', 'files' => true
                             ]) !!}
                                 <input name="_method" value="PUT" type="hidden" />
@@ -51,18 +51,18 @@
 
                                 <label>
                                     <input name="forum_cover[{{ $cover['key'] }}][_delete]" value="1" type="checkbox" />
-                                    {{ trans('forum.admin.forum-covers.delete') }}
+                                    {{ trans('admin.forum.forum-covers.delete') }}
                                 </label>
-                                <button>{{ trans('forum.admin.forum-covers.submit.update') }}</button>
+                                <button>{{ trans('admin.forum.forum-covers.submit.update') }}</button>
                             {!! Form::close() !!}
                         @else
-                            {{ trans('forum.admin.forum-covers.no-cover') }}
+                            {{ trans('admin.forum.forum-covers.no-cover') }}
 
-                            {!! Form::open(['url' => route('forum.admin.forum-covers.store'), 'method' => 'POST', 'files' => true]) !!}
+                            {!! Form::open(['url' => route('admin.forum.forum-covers.store'), 'method' => 'POST', 'files' => true]) !!}
                                 <input name="_method" value="POST" type="hidden" />
                                 <input name="forum_cover[forum_id]" value="{{ $forum->forum_id }}" type="hidden" />
                                 <input name="forum_cover[{{ $cover['key'] }}][cover_file]" type="file">
-                                <button>{{ trans('forum.admin.forum-covers.submit.save') }}</button>
+                                <button>{{ trans('admin.forum.forum-covers.submit.save') }}</button>
                             {!! Form::close() !!}
                         @endif
                     </div>
